pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weaver"
version = "0.1.0"
description = "Bidirectional context-weaving engine for AI agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Doug Allpress" }]
keywords = ["rag", "knowledge-graph", "ai-agents", "context"]

dependencies = [
    # Core CLI + secret store + HTTP + parsers (fast install; no torch)
    "click>=8.1",
    "rich>=13.7",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
    "keyring>=25.0",
    "cryptography>=42.0",
    "httpx>=0.27",
    "feedparser>=6.0",
    "wayfinder",
    "python-gitlab>=4.4",
    "GitPython>=3.1",
    "beautifulsoup4>=4.12",
    "lxml>=5.1",
    "markdown-it-py>=3.0",
    "defusedxml>=0.7",
    "charset-normalizer>=3.3",
    "ruamel.yaml>=0.18",
    "tomli-w>=1.0",
]

[project.optional-dependencies]
# Heavy, split out so a fresh install is quick for mail/auth work.
rag = ["chromadb>=0.5", "sentence-transformers>=2.7"]
graph = ["networkx>=3.2", "tree-sitter>=0.22", "tree-sitter-languages>=1.10"]
# Just networkx for the aggregator-built graph (no tree-sitter needed).
index = ["networkx>=3.2"]
pdf = ["pypdf>=4.2", "pdfplumber>=0.11"]
office = ["python-docx>=1.1", "python-pptx>=1.0", "python-calamine>=0.2", "openpyxl>=3.1"]
playwright = ["playwright>=1.42"]
dev = ["pytest>=8.1", "pytest-cov>=5.0", "ruff>=0.4", "mypy>=1.10"]

[project.scripts]
weaver = "weaver.cli.dispatcher:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["weaver*"]

[tool.setuptools.package-data]
"weaver" = ["py.typed", "**/*.ini", "**/*.toml", "**/*.yaml", "**/*.yml"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "ARG", "PTH"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-ra --strict-markers"
